Code cleanup for status reporting v1
