+++ services/gateway/app.js
@@ -30,7 +30,7 @@ const config = require('./config');
 const routes = require('./routes');
 const namespaces = require('./namespaces');
 const packageJson = require('./package.json');
-const { updateServiceStatus, getReady, getBuildTimestamp } = require('./shared/status');
+const { updateServiceStatus, getReady, getStatus } = require('./shared/status');
 const { genDocs } = require('./apis/http-version1/swagger/generateDocs');
 
 const mapper = require('./shared/customMapper');
@@ -69,24 +69,7 @@ broker.createService({
 	name: 'gateway',
 	actions: {
 		spec() { return genDocs(); },
-		async status() {
-			const buildTimestamp = getBuildTimestamp();
-			const networkstatus = await this.broker.call('core.network.status');
-			const networkStatistics = await this.broker.call('core.peers.statistics');
-			const { coreVer } = networkStatistics.data;
-			const versionCount = Object.values(coreVer);
-			const networkNodeVersion = Object.keys(coreVer)[
-				versionCount.indexOf(Math.max(...versionCount))
-			];
-			return {
-				build: buildTimestamp,
-				description: 'Lisk Service Gateway',
-				name: packageJson.name,
-				version: packageJson.version,
-				networkId: networkstatus.data.constants.nethash,
-				networkNodeVersion,
-			};
-		},
+		status() { return getStatus(this.broker); },
 		async ready() {
 			const services = await getReady();
 			// isReady: returns true if any one of service is unavailable
